Replace ping methods with remove
This makes it so that we can handle nodes dying at any point,
and not just the front of the bucket dying. This also allows
an arbitrary number of nodes to queue to replace elements of a bucket
when they die. Whenever we receive a ping response, we can just follow
the standard rule of inserting any node we receive a message from. We
can remove things if the rpc call fails to get a response regardless of
which rpc call it was in the first place.